requirements.txt
pyodbc>=4.0.39
python-dotenv>=1.0.0
pandas>=2.0.0
sqlalchemy>=2.0.0
requests>=2.31.0
